Split template for S3 bucket and S3 owner.txt
Sceptre has gotten smarter, and while we were previously able to launch
a bucket with the same stack name and not have the previous owner.txt
(from a different namespace) deleted, that's no longer the
case. Sceptre is actually tracking the custom CFN resource and removing
the previous instantiation when the namespace changes. This is not great
if we want to have a single bucket containing data for multiple
    namespaces, so I've split the bucket stack from the stack which
    creates the owner.txt.